Approx. conversion from linear RGB to sRGB color space
FF14 interprets material files in the sRGB color space. TexTools uses
linear RGB, which lead to colors appearing washed out when opened in
FF14. This implements an approximate conversion from linear RGB to sRGB,
making models in the 3D viewer and colors in the Colorset editor appear
more accurate to their in-game representation.